+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_equal_then_else/avoid_equal_then_else_rule.dart
@@ -37,11 +37,7 @@ class AvoidEqualThenElseRule extends CommonRule {
         .map(
           (node) => createIssue(
             rule: this,
-            location: nodeLocation(
-              node: node,
-              source: source,
-              withCommentOrMetadata: true,
-            ),
+            location: nodeLocation(node: node, source: source),
             message: _warningMessage,
           ),
         )