+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_conditional_expressions/prefer_conditional_expressions_rule.dart
@@ -40,11 +40,7 @@ class PreferConditionalExpressionsRule extends CommonRule {
         .map(
           (info) => createIssue(
             rule: this,
-            location: nodeLocation(
-              node: info.statement,
-              source: source,
-              withCommentOrMetadata: true,
-            ),
+            location: nodeLocation(node: info.statement, source: source),
             message: _warningMessage,
             replacement: _createReplacement(info),
           ),