+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_on_push_cd_strategy/prefer_on_push_cd_strategy_rule.dart
@@ -36,11 +36,7 @@ class PreferOnPushCdStrategyRule extends AngularRule {
     return visitor.expression
         .map((expression) => createIssue(
               rule: this,
-              location: nodeLocation(
-                node: expression,
-                source: source,
-                withCommentOrMetadata: true,
-              ),
+              location: nodeLocation(node: expression, source: source),
               message: _failure,
             ))
         .toList(growable: false);