+++ lib/src/analyzers/lint_analyzer/rules/rules_list/provide_correct_intl_args/provide_correct_intl_args_rule.dart
@@ -43,11 +43,7 @@ class ProvideCorrectIntlArgsRule extends IntlRule {
     return visitor.issues
         .map((issue) => createIssue(
               rule: this,
-              location: nodeLocation(
-                node: issue.node,
-                source: source,
-                withCommentOrMetadata: true,
-              ),
+              location: nodeLocation(node: issue.node, source: source),
               message: issue.nameFailure!,
             ))
         .toList();